Improved initialization of sync/unsync buttons
To change the icon of the sync/unsync button based on player status, that button object needs to be initialized in the setup section of button_action_functions.py. Previously the address was hard coded there instead of taking it as an input from button_defs.json. This change puts everything into button_defs.json.