Major rework of clock setting
Striped out code that went to worldtimeapi.org to get time zone, DST,  and current time.
Instead, uses ntptime and a local tz data file to set clock  and to keep up with when to change to/from DTS.
Also include the tz data file and the code to regenerate that file.